+++ test/mocha/10-keystore.js
@@ -3,8 +3,12 @@
  */
 'use strict';
 
+const bedrock = require('bedrock');
 const {CapabilityAgent} = require('webkms-client');
 const helpers = require('./helpers');
+const brHttpsAgent = require('bedrock-https-agent');
+const axios = require('axios');
+const mockData = require('./mock.data');
 
 describe('bedrock-kms-http API', () => {
   describe('keystores', () => {
@@ -31,6 +35,204 @@ describe('bedrock-kms-http API', () => {
       result.should.have.property('controller');
       result.controller.should.equal(capabilityAgentId);
     });
+    it('throws error on no sequence in postKeystore validation', async () => {
+      const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+      const handle = 'testKey1';
+
+      const capabilityAgent = await CapabilityAgent
+        .fromSecret({secret, handle});
+      const kmsBaseUrl = `${bedrock.config.server.baseUri}/kms`;
+      const url = `${kmsBaseUrl}/keystores`;
+      const config = {
+        controller: capabilityAgent.id
+      };
+      const {httpsAgent} = brHttpsAgent;
+
+      let err;
+      let result;
+      try {
+        result = await axios.post(
+          url,
+          config,
+          {httpsAgent}
+        );
+      } catch(e) {
+        err = e;
+      }
+      should.exist(err);
+      should.not.exist(result);
+      err.response.data.message.should.equal(
+        'A validation error occured in the \'postKeystore\' validator.');
+    });
+    it('gets a keystore', async () => {
+      const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+      const handle = 'testKey1';
+
+      const capabilityAgent = await CapabilityAgent
+        .fromSecret({secret, handle});
+
+      const keystore = await helpers.createKeystore({capabilityAgent});
+      let err;
+      let result;
+      try {
+        result = await helpers.getKeystore({id: keystore.id});
+      } catch(e) {
+        err = e;
+      }
+      assertNoError(err);
+      should.exist(result);
+      result.should.have.property('id');
+      result.id.should.equal(keystore.id);
+    });
+    it('finds a keystore', async () => {
+      const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+      const handle = 'testKey1';
+      // eslint-disable-next-line max-len
+      const referenceId = 'did:key:z6MkkrtV7wnBpXKBtiZjxaSghCo8ttb5kZUJTk8bEwTTTYvg';
+
+      const capabilityAgent = await CapabilityAgent
+        .fromSecret({secret, handle});
+
+      const keystore = await helpers.createKeystore({
+        capabilityAgent, referenceId});
+      let err;
+      let result;
+      try {
+        result = await helpers.findKeystore({
+          controller: keystore.controller, referenceId});
+      } catch(e) {
+        err = e;
+      }
+      assertNoError(err);
+      should.exist(result);
+      result.should.have.property('id');
+      result.id.should.equal(keystore.id);
+      result.controller.should.equal(keystore.controller);
+      result.referenceId.should.equal(keystore.referenceId);
+    });
+    it('throws error on no controller in findKeystore validation', async () => {
+      const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+      const handle = 'testKey1';
+      // eslint-disable-next-line max-len
+      const referenceId = 'did:key:z6MkkrtV7wnBpXKBtiZjxaSghCo8ttb5kZUJTk8bEwTTTYvg';
+
+      const capabilityAgent = await CapabilityAgent
+        .fromSecret({secret, handle});
+
+      const keystore = await helpers.createKeystore({
+        capabilityAgent, referenceId});
+
+      const kmsBaseUrl = `${bedrock.config.server.baseUri}/kms`;
+      const url = `${kmsBaseUrl}/keystores` +
+      `/?r?controller=${keystore.controller}`;
+
+      const {httpsAgent} = brHttpsAgent;
+
+      let err;
+      let result;
+      try {
+        result = await axios.get(
+          url,
+          {httpsAgent}
+        );
+      } catch(e) {
+        err = e;
+      }
+      should.exist(err);
+      should.not.exist(result);
+      err.response.data.message.should.equal(
+        'A validation error occured in the \'findKeystore\' validator.');
+    });
+    it('throws error on no referenceId in findKeystore validation',
+      async () => {
+      // eslint-disable-next-line max-len
+        const referenceId = 'did:key:z6MkkrtV7wnBpXKBtiZjxaSghCo8ttb5kZUJTk8bEwTTTYvg';
+
+        const kmsBaseUrl = `${bedrock.config.server.baseUri}/kms`;
+        const url = `${kmsBaseUrl}/keystores` +
+      `/?referenceId=${referenceId}`;
+
+        const {httpsAgent} = brHttpsAgent;
+
+        let err;
+        let result;
+        try {
+          result = await axios.get(
+            url,
+            {httpsAgent}
+          );
+        } catch(e) {
+          err = e;
+        }
+        should.exist(err);
+        should.not.exist(result);
+        err.response.data.message.should.equal(
+          'A validation error occured in the \'findKeystore\' validator.');
+      });
+    it('throws error with no invoker in zcap validation', async () => {
+      const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+      const handle = 'testKey1';
+
+      const capabilityAgent = await CapabilityAgent
+        .fromSecret({secret, handle});
+      const keystore = await helpers.createKeystore({
+        capabilityAgent});
+
+      const url = `${keystore.id}/authorizations`;
+
+      const zcap = mockData.zcaps.zero;
+      delete zcap.invoker;
+      const {httpsAgent} = brHttpsAgent;
+
+      let err;
+      let result;
+      try {
+        result = await axios.post(
+          url,
+          zcap,
+          {httpsAgent}
+        );
+      } catch(e) {
+        err = e;
+      }
+      should.exist(err);
+      should.not.exist(result);
+      err.response.data.message.should.equal(
+        'A validation error occured in the \'zcap\' validator.');
+    });
+    it('throws error with no controller in recovery validation',
+      async () => {
+        const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
+        const handle = 'testKey1';
+
+        const capabilityAgent = await CapabilityAgent
+          .fromSecret({secret, handle});
+        const keystore = await helpers.createKeystore({
+          capabilityAgent});
+
+        const url = `${keystore.id}/recover`;
+
+        const config = {
+          '@context': 'https://w3id.org/security/v2',
+        };
+        const {httpsAgent} = brHttpsAgent;
+
+        let err;
+        let result;
+        try {
+          result = await axios.post(
+            url,
+            config,
+            {httpsAgent}
+          );
+        } catch(e) {
+          err = e;
+        }
+        should.exist(err);
+        should.not.exist(result);
+        err.response.data.message.should.equal(
+          'A validation error occured in the \'recovery\' validator.');
+      });
     it('throws error on receivedHost not equal to allowedHost', async () => {
       const secret = ' b07e6b31-d910-438e-9a5f-08d945a5f676';
       const handle = 'testKey1';